--- src/main/java/ru/kata/spring/boot_security/demo/services/UserDetailsServiceImp.java
@@ -1,31 +1,0 @@
-package ru.kata.spring.boot_security.demo.services;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import ru.kata.spring.boot_security.demo.models.User;
-import ru.kata.spring.boot_security.demo.repositories.UserRepository;
-
-@Service
-public class UserDetailsServiceImp implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    @Autowired
-    public UserDetailsServiceImp(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Transactional
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        User user = userRepository.findByUsername(username);
-        if (user == null) {
-            throw new UsernameNotFoundException("User not found");
-        }
-        return user;
-    }
-}
